+++ currenapp/main.py
@@ -1,46 +1,62 @@
+import re
 import typer
 import requests
 import os
 from dotenv import load_dotenv
-from rich import print
+from rich import print as rprint
+from rich.console import Console
+
+console = Console()
+
 
 load_dotenv()
-app = typer.Typer()
+# Disable local variables on error logs on production
+app = typer.Typer(
+    pretty_exceptions_show_locals=os.getenv("SHOW_LOCALS", False))
 
 
-@app.callback()
-def callback():
-    # some Documentation
-    """
-    Curruncy Conversion
+def get_apidata(amount: float, base: str, target: str) -> dict:
+    API_KEY = os.getenv("API_KEY")
+    url = f"https://v6.exchangerate-api.com/v6/{API_KEY}/pair/{base}/{target}/{amount}"
 
-    Converts a passed VALUE from --BASE to --TARGET currency using  the exchangerate-api.
-    """
+    response = requests.get(url)
+    return response.json()
+
+
+def is_validargs(base: str, target: str) -> bool:
+    # check for valid currency code
+    # Note: Only checking if the input is on the right format (3 uppercase letters from A-Z)
+    pattern = r"^[A-Z]{3}$"
+    return re.match(pattern, base) and re.match(pattern, target)
 
 
 @app.command()
-def main(value: float, base: str = "USD", target: str = "BRL"):
-    if len(base) != 3 or len(target) != 3:
-        print(
+def main(base: str = "USD", target: str = "BRL"):
+    if not is_validargs(base, target):
+        rprint(
             f"[bold red]Invalid currency codes![/bold red] Received: {base}, {target}. Check your inputs.")
         raise typer.Abort()
 
-    API_KEY = os.getenv("API_KEY")
-    url = f"https://v6.exchangerate-api.com/v6/{API_KEY}/pair/{base}/{target}/{value}"
+    try:
+        amount = float(typer.prompt("Value to convert $"))
+    except ValueError:
+        rprint("[bold red]Expected a number![/bold red].")
+        raise typer.Abort()
 
-    response = requests.get(url)
-    data: dict = response.json()
+    with console.status("Making the conversion", spinner="bouncingBall"):
+        data = get_apidata(amount, base, target)
 
     result = data.get("result")
     match result:
         case "success":
-            print("Conversion")
-            print(f"${base}: {value} = ${target}: {data.get('conversion_result')}")
-            print(f"Conversion rate: {data.get('conversion_rate')}")
-            print(f"Next Update in {data.get('time_next_update_utc')}")
+            rprint("Conversion")
+            rprint(
+                f"${base}: {amount} = ${target}: {data.get('conversion_result')}")
+            rprint(f"Conversion rate: {data.get('conversion_rate')}")
+            rprint(f"Next Update in {data.get('time_next_update_utc')}")
 
         case "error":
-            print("Something went wrong")
+            rprint("Something went wrong...")
             print(f"Error Type: {data.get('error_type')}")
             typer.Exit(code=1)