+++ CHANGELOG.md
@@ -8,13 +8,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - [ ] Full Documentation
-- [ ] Unit tests
-- [ ] Better error handling
-- [ ] Hide local variables
 - [ ] Beautiful CLI
 - [ ] Offering PiPy installation
 
-## 0.1.0 - 2022-09-28
+## 0.2.0 — 2022-10-01
+
+### Added
+
+- Tests on tests/test_main.py
+- Status bar for the conversion request
+- Improve validation of currency code.
+- HTTMock as a dependencie for tests.
+
+### Modified
+
+- Value to be converted comes from a prompt question instead of a argument.
+- Move validation and request to the API to their own functions.
+- Validation of currency code now uses RegEx.
+
+### Fixed
+
+- No more calling "main" as a command to use the application.
+
+### Removed
+
+- Callback with documetation, since it was causing the problem with "main".
+
+### Security
+
+- Hide local variables on production.
+
+## 0.1.0 — 2022-09-28
 
 ### Added